新来的成员欢迎bug

+++ telegram/telegram.go
@@ -184,7 +184,7 @@ func (h *Telegram) Forward(msg *tgbotapi.Message) {
 	if botMsg != nil {
 		if msg.NewChatMember != nil {
 			botMsg.From = msg.NewChatMember.UserName
-			botMsg.UID = msg.NewChatMember.ID
+			botMsg.UID = int64(msg.NewChatMember.ID)
 		} else {
 			botMsg.From = msg.From.UserName
 			botMsg.UID = int64(msg.From.ID)
